high and low level clients can deal with string errors
