+++ src/Tests/ClientConcepts/ServerError/ComplexErrorTests.cs
@@ -1,0 +1,112 @@
+﻿using Elasticsearch.Net;
+using FluentAssertions;
+using Tests.Framework;
+
+namespace Tests.ClientConcepts.ServerError
+{
+	public class ComplexErrorTests : ServerErrorTestsBase
+	{
+		[U] protected override void AssertServerError() => base.AssertServerError();
+
+		protected override string Json  =>@"{
+	""root_cause"" : [
+	{
+		""type"" : ""parse_exception"",
+		""reason"" : ""failed to parse date field [-1m] with format [strict_date_optional_time||epoch_millis]""
+	}],
+	""type"" : ""search_phase_execution_exception"",
+	""reason"" : ""all shards failed"",
+	""phase"" : ""query"",
+	""grouped"" : true,
+	""failed_shards"" : [
+	{
+		""shard"" : 0,
+		""index"" : ""project"",
+		""node"" : ""Uo6PBln_QrmD8Y9o1NKdQw"",
+		""reason"" : {
+			""type"" : ""parse_exception"",
+			""reason"" : ""failed to parse date field [-1m] with format [strict_date_optional_time||epoch_millis]"",
+			""caused_by"" : {
+				""type"" : ""illegal_argument_exception"",
+				""reason"" : ""Parse failure at index [2] of [-1m]""
+			}
+		}
+	}
+	],
+	""headers"" : {
+		""WWW-Authenticate"" : ""Bearer: ..."",
+		""x"" : ""y""
+	},
+	""caused_by"" : {
+		""type"" : ""parse_exception"",
+		""reason"" : ""failed to parse date field [-1m] with format [strict_date_optional_time||epoch_millis]"",
+		""index"" : null,
+		""resource.id"" : [""alias1"", ""alias2""],
+		""script_stack"" : [""alias1"", ""alias2""],
+		""unknown_prop"" : [""alias1"", ""alias2""],
+		""caused_by"" : {
+			""type"" : ""illegal_argument_exception"",
+			""reason"" : ""Parse failure at index [2] of [-1m]"",
+			""caused_by"" : ""x""
+		}
+	},
+	""license.expired.feature"" : ""ml"",
+	""index"" : ""index"",
+	""index_uuid"" : ""x9h1ks"",
+	""unknown_prop"" : {},
+	""unknown_prop2"" : false,
+	""resource.type"" : ""aliases"",
+	""resource.id"" : ""alias1"",
+	""shard"" : ""1"",
+	""line"" : 12,
+	""col"" : 199,
+	""bytes_wanted"" : 1298312,
+	""bytes_limit"" : 8912031,
+	""script_stack"" : ""x"",
+	""script"" : ""some script"",
+	""lang"" : ""c#""
+}";
+
+		protected override void AssertResponseError(string origin, Error error)
+		{
+			AssertCausedBy(origin, error);
+			AssertCausedBy(origin, error.CausedBy);
+			AssertCausedBy(origin, error.CausedBy.CausedBy);
+			error.CausedBy.CausedBy.CausedBy.Should().NotBeNull();
+			error.CausedBy.CausedBy.CausedBy.Reason.Should().Be("x");
+			error.RootCause.Should().NotBeEmpty(origin);
+			error.Headers.Should().HaveCount(2, origin);
+			AssertMetadata(origin, error.Metadata);
+			error.CausedBy.Metadata.Should().NotBeNull();
+			error.CausedBy.Metadata.ScriptStack.Should().HaveCount(2);
+			error.CausedBy.Metadata.ResourceId.Should().HaveCount(2);
+		}
+
+		private void AssertMetadata(string origin, ErrorCause.ErrorCauseMetadata errorMetadata)
+		{
+			errorMetadata.Should().NotBeNull(origin);
+			errorMetadata.Grouped.Should().BeTrue(origin);
+			errorMetadata.Phase.Should().Be("query", origin);
+			errorMetadata.LicensedExpiredFeature.Should().Be("ml", origin);
+			errorMetadata.Index.Should().Be("index", origin);
+			errorMetadata.IndexUUID.Should().NotBeNullOrWhiteSpace(origin);
+			errorMetadata.ResourceType.Should().NotBeNullOrWhiteSpace(origin);
+			errorMetadata.ResourceId.Should().HaveCount(1, origin);
+			errorMetadata.Shard.Should().Be(1, origin);
+			errorMetadata.Line.Should().Be(12, origin);
+			errorMetadata.Column.Should().Be(199, origin);
+			errorMetadata.BytesWanted.Should().BeGreaterThan(1, origin);
+			errorMetadata.BytesLimit.Should().BeGreaterThan(1, origin);
+			errorMetadata.ScriptStack.Should().HaveCount(1, origin);
+			errorMetadata.Script.Should().NotBeNullOrWhiteSpace(origin);
+			errorMetadata.Language.Should().NotBeNullOrWhiteSpace(origin);
+		}
+
+		private static void AssertCausedBy(string origin, ErrorCause causedBy)
+		{
+			causedBy.Should().NotBeNull(origin);
+			causedBy.Type.Should().NotBeNullOrWhiteSpace(origin);
+			causedBy.Reason.Should().NotBeNullOrWhiteSpace(origin);
+		}
+	}
+}
